Restructuring a bit.
Adding a code debug_only method.
Re-shuffling how we init/free svgs.
Planning for some 'comment skeletons' to say where new dev stuff should go.
Updated git ignore to ignore CMake stuff for Clion

+++ compute/src/main.h
@@ -9,8 +9,8 @@ typedef struct Config {
     char* inputFilePath;
     char* outputFilePath;
     char** extra_args;
-    int width;
-    int height;
+    long width;
+    long height;
 } Config;
 
 #include <stdio.h>
